Remove dust_api.ts confusing file and move methods to hooks
