+++ front/lib/swr/assistants.ts
@@ -1,10 +1,12 @@
 import type {
+  AgentConfigurationScope,
   AgentConfigurationType,
   AgentsGetViewType,
+  AgentUserListStatus,
   LightAgentConfigurationType,
   LightWorkspaceType,
 } from "@dust-tt/types";
-import { useContext, useMemo } from "react";
+import { useCallback, useContext, useMemo } from "react";
 import type { Fetcher } from "swr";
 
 import { SendNotificationsContext } from "@app/components/sparkle/Notification";
@@ -14,10 +16,12 @@ import {
   useSWRWithDefaults,
 } from "@app/lib/swr/swr";
 import type { GetAgentConfigurationsResponseBody } from "@app/pages/api/w/[wId]/assistant/agent_configurations";
+import type { PostAgentScopeRequestBody } from "@app/pages/api/w/[wId]/assistant/agent_configurations/[aId]/scope";
 import type { GetAgentUsageResponseBody } from "@app/pages/api/w/[wId]/assistant/agent_configurations/[aId]/usage";
 import type { GetSlackChannelsLinkedWithAgentResponseBody } from "@app/pages/api/w/[wId]/assistant/builder/slack/channels_linked_with_agent";
 import type { FetchAssistantTemplatesResponse } from "@app/pages/api/w/[wId]/assistant/builder/templates";
 import type { FetchAssistantTemplateResponse } from "@app/pages/api/w/[wId]/assistant/builder/templates/[tId]";
+import type { PostAgentListStatusRequestBody } from "@app/pages/api/w/[wId]/members/me/agent_list_status";
 
 export function useAssistantTemplates({
   workspaceId,
@@ -316,4 +320,167 @@ export function useDeleteAgentConfiguration({
   };
 
   return doDelete;
+}
+
+export function useUpdateAgentScope({
+  owner,
+  agentConfigurationId,
+}: {
+  owner: LightWorkspaceType;
+  agentConfigurationId: string | null;
+}) {
+  const sendNotification = useContext(SendNotificationsContext);
+  const { mutateAgentConfiguration: mutateCurrentAgentConfiguration } =
+    useAgentConfiguration({
+      workspaceId: owner.sId,
+      agentConfigurationId,
+      disabled: true,
+    });
+  const { mutate: mutateAgentConfigurations } =
+    useProgressiveAgentConfigurations({
+      workspaceId: owner.sId,
+      disabled: true,
+    });
+
+  const doUpdate = useCallback(
+    async (scope: Exclude<AgentConfigurationScope, "global">) => {
+      const body: PostAgentScopeRequestBody = {
+        scope,
+      };
+
+      try {
+        if (!agentConfigurationId) {
+          throw new Error(
+            "Cannot update scope of a non-existing agent. Action: make sure agentConfigurationId is not null."
+          );
+        }
+
+        const res = await fetch(
+          `/api/w/${owner.sId}/assistant/agent_configurations/${agentConfigurationId}/scope`,
+          {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify(body),
+          }
+        );
+
+        if (res.ok) {
+          sendNotification({
+            title: `Assistant sharing updated.`,
+            type: "success",
+          });
+          await mutateAgentConfigurations();
+          await mutateCurrentAgentConfiguration();
+          return true;
+        } else {
+          const data = await res.json();
+          sendNotification({
+            title: `Error updating assistant sharing.`,
+            description: data.error.message,
+            type: "error",
+          });
+          return false;
+        }
+      } catch (error) {
+        sendNotification({
+          title: `Error updating assistant sharing.`,
+          description: (error as Error).message || "An unknown error occurred",
+          type: "error",
+        });
+        return false;
+      }
+    },
+    [
+      agentConfigurationId,
+      mutateAgentConfigurations,
+      mutateCurrentAgentConfiguration,
+      owner.sId,
+      sendNotification,
+    ]
+  );
+  return doUpdate;
+}
+
+export function useUpdateAgentUserListStatus({
+  owner,
+  agentConfigurationId,
+}: {
+  owner: LightWorkspaceType;
+  agentConfigurationId: string;
+}) {
+  const sendNotification = useContext(SendNotificationsContext);
+  const { mutateAgentConfiguration: mutateCurrentAgentConfiguration } =
+    useAgentConfiguration({
+      workspaceId: owner.sId,
+      agentConfigurationId,
+      disabled: true,
+    });
+  const { mutate: mutateAgentConfigurations } =
+    useProgressiveAgentConfigurations({
+      workspaceId: owner.sId,
+      disabled: true,
+    });
+
+  const doUpdate = useCallback(
+    async (listStatus: AgentUserListStatus) => {
+      try {
+        const body: PostAgentListStatusRequestBody = {
+          agentId: agentConfigurationId,
+          listStatus,
+        };
+
+        const res = await fetch(
+          `/api/w/${owner.sId}/members/me/agent_list_status`,
+          {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify(body),
+          }
+        );
+
+        if (res.ok) {
+          sendNotification({
+            title: `Assistant ${
+              listStatus === "in-list"
+                ? "added to your list"
+                : "removed from your list"
+            }`,
+            type: "success",
+          });
+          await mutateAgentConfigurations();
+          await mutateCurrentAgentConfiguration();
+          return true;
+        } else {
+          const data = await res.json();
+          sendNotification({
+            title: `Error ${
+              listStatus === "in-list" ? "adding" : "removing"
+            } Assistant`,
+            description: data.error.message,
+            type: "error",
+          });
+          return false;
+        }
+      } catch (error) {
+        sendNotification({
+          title: `Error updating assistant list.`,
+          description: (error as Error).message || "An unknown error occurred",
+          type: "error",
+        });
+        return false;
+      }
+    },
+    [
+      agentConfigurationId,
+      mutateAgentConfigurations,
+      mutateCurrentAgentConfiguration,
+      owner.sId,
+      sendNotification,
+    ]
+  );
+  return doUpdate;
 }
